Move DependencyMap to its own module.
